move Brightness from System/ to Resource/
