+++ config/usa/arm9/delinks.txt
@@ -69,6 +69,10 @@ src/World/Object3D.cpp:
     .data start:0x020efaa8 end:0x020efb44
     .bss start:0x02104b18 end:0x02104b60
 
+src/Resource/Brightness.cpp:
+    complete
+    .text start:0x0203aee0 end:0x0203b498
+
 src/Graphics/LightingManager.cpp:
     complete
     .text start:0x02050454 end:0x02052958
@@ -154,10 +158,6 @@ src/Grotto/Main/DetailedTreasureMapData.cpp:
 src/Grotto/Main/TreasureMapMetadata.cpp:
     complete
     .text start:0x020a5cb8 end:0x020a6084
-
-src/System/Brightness.cpp:
-    complete
-     .text start:0x0203aee0 end:0x0203b498
 
 src/Memory/SignedAllocator.cpp:
     complete
