+++ include/Resource/Brightness.h
